started refactoring the project and code
