+++ client/src/components/DailyDevotion/Devotion.jsx
@@ -44,7 +44,7 @@ const Work = () => {
           <div
             key={index}
             onClick={() => handleDevotionFilter(item)}
-            className={`app__work-filter-item app__flex p-text ${activeFilter === item ? 
+            className={`app__work-filter-item app__flex p-text ${activeFilter === item ?
               'item-active' : ''}`}
           >
             {item}
@@ -69,8 +69,8 @@ const Work = () => {
                 transition={{ duration: 0.25, ease: 'easeInOut', staggerChildren: 0.5 }}
                 className="app__work-hover app__flex"
               >
-                <a href={work.projectLink} 
-                target="_blank" rel="noreferrer">
+                <a href={work.projectLink}
+                  target="_blank" rel="noreferrer">
 
                   <motion.div
                     whileInView={{ scale: [0, 1] }}
