+++ client/src/components/Gallery/Gallery.jsx
@@ -2,14 +2,14 @@ import { useState } from 'react'
 
 import { FontAwesomeIcon } from '@fortawesome/react-fontawesome'
 import {
-  faCircleChevronLeft, 
-  faCircleChevronRight, 
+  faCircleChevronLeft,
+  faCircleChevronRight,
   faCircleXmark
 } from '@fortawesome/free-solid-svg-icons'
 
 import './Gallery.css'
 
-const Gallery = ({galleryImages}) => {
+const Gallery = ({ galleryImages }) => {
 
   const [slideNumber, setSlideNumber] = useState(0)
   const [openModal, setOpenModal] = useState(false)
@@ -26,24 +26,24 @@ const Gallery = ({galleryImages}) => {
 
   // Previous Image
   const prevSlide = () => {
-    slideNumber === 0 
-    ? setSlideNumber( galleryImages.length -1 ) 
-    : setSlideNumber( slideNumber - 1 )
+    slideNumber === 0
+      ? setSlideNumber(galleryImages.length - 1)
+      : setSlideNumber(slideNumber - 1)
   }
 
   // Next Image  
   const nextSlide = () => {
-    slideNumber + 1 === galleryImages.length 
-    ? setSlideNumber(0) 
-    : setSlideNumber(slideNumber + 1)
+    slideNumber + 1 === galleryImages.length
+      ? setSlideNumber(0)
+      : setSlideNumber(slideNumber + 1)
   }
 
   return (
     <div>
 
-      {openModal && 
+      {openModal &&
         <div className='sliderWrap'>
-          <FontAwesomeIcon icon={faCircleXmark} className='btnClose' onClick={handleCloseModal}  />
+          <FontAwesomeIcon icon={faCircleXmark} className='btnClose' onClick={handleCloseModal} />
           <FontAwesomeIcon icon={faCircleChevronLeft} className='btnPrev' onClick={prevSlide} />
           <FontAwesomeIcon icon={faCircleChevronRight} className='btnNext' onClick={nextSlide} />
           <div className='fullScreenImage'>
@@ -52,20 +52,14 @@ const Gallery = ({galleryImages}) => {
         </div>
       }
 
-      {/* <br />
-      Current slide number:  {slideNumber}
-      <br />
-      Total Slides: {galleryImages.length}
-      <br /><br /> */}
-
       <div className='galleryWrap'>
         {
           galleryImages && galleryImages.map((slide, index) => {
-            return(
-              <div 
-                className='single' 
+            return (
+              <div
+                className='single'
                 key={index}
-                onClick={ () => handleOpenModal(index) }
+                onClick={() => handleOpenModal(index)}
               >
                 <img src={slide.img} alt='' />
               </div>
