--- client/src/components/Header/Header.jsx
@@ -1,24 +1,0 @@
-import React from 'react';
-import { motion } from 'framer-motion';
-import { AppWrap } from '../../wrapper';
-import { images } from '../../constants';
-import './Header.scss';
-import Slider from '../Slider/slider';
-const scaleVariants = {
-  whileInView: {
-    scale: [0, 1],
-    opacity: [0, 1],
-    transition: {
-      duration: 1,
-      ease: 'easeInOut',
-    },
-  },
-};
-
-const Header = () => (
-  <div className="">
-{/* <Slider /> */}
-  </div>
-);
-
-export default AppWrap(Header, 'home');
