+++ client/src/components/Header/Header.scss
@@ -15,26 +15,7 @@
     }
   }
   
-  // .app__header {
-  //   flex: 1;
-  //   width: 100%;
-  //   height: 100%;
-  //   flex-direction: row;
-  
-  //   padding: 6rem 2rem 0rem;
-  
-  //   @media screen and (min-width: 2000px) {
-  //     padding-top: 8rem;
-  //   }
-  
-  //   @media screen and (max-width: 1200px) {
-  //     flex-direction: column;
-  //   }
-  
-  //   @media screen and (max-width: 450px) {
-  //     padding: 6rem 1rem 2rem;
-  //   }
-  // }
+
   
   .app__header-info {
     flex: 0.65;
