+++ client/src/components/Navbar/Navbar.jsx
@@ -1,5 +1,4 @@
 import React from 'react';
-import { Link } from 'react-router-dom';
 import './Navbar.css';
 import Header from './Header'
 function Navbar() {
@@ -8,7 +7,6 @@ function Navbar() {
       <Header />
     </div>
   );
-
 }
 
 export default Navbar
