+++ client/src/components/Testimonials/Testimonials.jsx
@@ -1,6 +1,5 @@
 import React, { useState, useEffect } from 'react';
 import { HiChevronLeft, HiChevronRight } from 'react-icons/hi';
-import { motion } from 'framer-motion';
 
 import { AppWrap, MotionWrap } from '../../wrapper';
 import { urlFor, client } from '../../client';
@@ -9,7 +8,6 @@ import './Testimonials.scss';
 const Testimonial = () => {
   const [currentIndex, setCurrentIndex] = useState(0);
   const [testimonials, setTestimonials] = useState([]);
-  const [brands, setBrands] = useState([]);
 
   const handleClick = (index) => {
     setCurrentIndex(index);
@@ -26,24 +24,24 @@ const Testimonial = () => {
 
   return (
     <>
-     <h2 className="head-text">Testimonials</h2>
+      <h2 className="head-text">Testimonials</h2>
       {testimonials.length && (
         <>
           <div className="app__testimonial-item app__flex">
-  <img src={urlFor(testimonials[currentIndex].imageUrl)} 
-  alt={testimonials[currentIndex].name} />
+            <img src={urlFor(testimonials[currentIndex].imageUrl)}
+              alt={testimonials[currentIndex].name} />
             <div className="app__testimonial-content">
               <p className="p-text">{testimonials[currentIndex].feedback}</p>
               <div>
-      <h4 className="bold-text">{testimonials[currentIndex].name}</h4>
-      <h5 className="p-text">{testimonials[currentIndex].organisation}</h5>
+                <h4 className="bold-text">{testimonials[currentIndex].name}</h4>
+                <h5 className="p-text">{testimonials[currentIndex].organisation}</h5>
               </div>
             </div>
           </div>
 
           <div className="app__testimonial-btns app__flex">
-            <div className="app__flex" onClick={() => 
-            handleClick(currentIndex === 0 ? testimonials.length - 1 : currentIndex - 1)}>
+            <div className="app__flex" onClick={() =>
+              handleClick(currentIndex === 0 ? testimonials.length - 1 : currentIndex - 1)}>
               <HiChevronLeft />
             </div>
