+++ client/src/constants/images.js
@@ -1,12 +1,18 @@
-import email from '../assets/email.png';
-import mobile from '../assets/mobile.png';
-import profile from '../assets/profile.png';
-import circle from '../assets/circle.svg';
-import bropic from '../assets/bro pic.jpg';
-import mypic from '../assets/my profile.jpg';
-import healthpicone from '../assets/Health pic.jpeg';
-import healthpictwo from '../assets/healthpic2.jpeg';
-import healthpicthree from '../assets/Health pic3.jpeg';
+import email from "../assets/email.png";
+import mobile from "../assets/mobile.png";
+import profile from "../assets/profile.png";
+import circle from "../assets/circle.svg";
+import bropic from "../assets/bro pic.jpg";
+import mypic from "../assets/my profile.jpg";
+import healthpicone from "../assets/Health pic.jpeg";
+import healthpictwo from "../assets/healthpic2.jpeg";
+import healthpicthree from "../assets/Health pic3.jpeg";
+import hannah from "../assets/hannah.jpg";
+import salaton from "../assets/Salaton.jpg";
+import pastorKyuSda from "../assets/pastorkyusda.jpeg";
+import elderSam from "../assets/Eldersam.jpg";
+import joshua from "../assets/Joshua pic.jpg";
+import ibra from "../assets/ibra.jpg";
 export default {
   email,
   mobile,
@@ -16,5 +22,11 @@ export default {
   mypic,
   healthpicone,
   healthpictwo,
-  healthpicthree
+  healthpicthree,
+  ibra,
+  hannah,
+  salaton,
+  pastorKyuSda,
+  elderSam,
+  joshua,
 };
