+++ client/src/container/index.js
@@ -1,14 +1,10 @@
-import { FormGroup } from '@material-ui/core';
-
-export {default as About } from '../components/About/About';
-export {default as Header} from '../components/Header/Header';
-export {default as Testimonial} from
- "../components/Testimonials/Testimonials"
-export {default as Departments} from "../components/Work/Departments";
-export {default as Families} from "../components/Family/Families";
-export {default as Gallery} from "../components/ImageGallery/ImageGallery"
-export {default as Newmember} from "../components/Member"
-export {default as Announcements} from "../components/Announcements/Announcement.jsx"
-export {default as Footer} from '../components/Footer/Footer.jsx'
-export {default as Kyusda} from '../components/Gallery/Kyusdagallery';
-export {default as Leader} from '../components/Leaders';
+export { default as About } from "../components/About/About";
+export { default as Testimonial } from "../components/Testimonials/Testimonials";
+export { default as Departments } from "../components/Work/Departments";
+export { default as Families } from "../components/Family/Families";
+export { default as Gallery } from "../components/ImageGallery/ImageGallery";
+export { default as NewMember } from "../components/Member";
+export { default as Announcements } from "../components/Announcements/Announcement.jsx";
+export { default as Footer } from "../components/Footer/Footer.jsx";
+export { default as KyuSda } from "../components/Gallery/KyuSdaGallery.js";
+export { default as Leader } from "../components/Leaders";
